ci: rework workflows to be triggered by Ariane on 1.11
In this commit, we move stable workflows from the `main` branch back
into the 1.11 stable branch now that workflows are triggered via
`workflow_dispatch` events in the appropriate context.

The full details around this new mechanism are documented in the commit
made on `main`: <TBD>

Since these new workflows were previously living in `main`, we also need
to backport dependencies on `.github/actions` configuration files. We
take the opportunity to adjust the configuration files as appropriate,
notably in terms of K8s version coverage, to ensure that we only test
K8s versions officially supported by the stable branch.

In particular for 1.11, the AKS workflow was NOT backported because 1.11
only supports K8s versions up to 1.23, and 1.23 is not available on AKS
anymore. The stable 1.11 AKS workflow from `main` was already disabled
when this changed happened, as per our testing policy, so we are not
backporting it.

Signed-off-by: Nicolas Busseneau <nicolas@isovalent.com>