ci: rework workflows to be triggered by Ariane on 1.12
In this commit, we move stable workflows from the `main` branch back
into the 1.12 stable branch now that workflows are triggered via
`workflow_dispatch` events in the appropriate context.

The full details around this new mechanism are documented in the commit
made on `main`: <TBD>

Since these new workflows were previously living in `main`, we also need
to backport dependencies on `.github/actions` configuration files. We
take the opportunity to adjust the configuration files as appropriate,
notably in terms of K8s version coverage, to ensure that we only test
K8s versions officially supported by the stable branch.

Signed-off-by: Nicolas Busseneau <nicolas@isovalent.com>